requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
lark-oapi==1.2.6
requests==2.31.0

# 数据库和缓存
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# 安全和认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 异步和任务队列
celery==5.3.4
flower==2.0.1
httpx==0.25.1

# 监控和日志
prometheus-fastapi-instrumentator==5.9.1
structlog==23.2.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
